# backend/src/classrooms/classrooms.controller.ts
import {
  Request,
  Body,
  Controller,
  HttpCode,
  HttpStatus,
  Post,
  UseGuards,
  Param,
  Get,
  Query,
  Delete,
} from '@nestjs/common';
import {
  ApiBearerAuth,
  ApiBody,
  ApiExtraModels,
  ApiOperation,
  ApiParam,
  ApiQuery,
  ApiResponse,
  ApiTags,
} from '@nestjs/swagger';
import { RolesGuard } from '../auth/guards/roles.guard';
import { Role } from '@prisma/client';
import { AuthUser } from '../common/types/auth-user.type';
import { CreateClassroomDto } from './dto/create-classroom.dto';
import { FindAvailableClassroomsQueryDto } from './dto/find-available-classrooms-query.dto';
import {
  ClassroomMemberDto,
  ClassroomSummaryDto,
  ClassroomWithMembersDto,
  LastAnnouncementSummaryDto,
  TeacherSummaryDto,
} from './dto/classroom-summary.dto';
import { ClassroomsService } from './classrooms.service';
import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
import { Roles } from '../auth/decorator/roles.decorator';
import {
  ApiConflictResponse,
  ApiForbiddenResponse,
  ApiNotFoundResponse,
  ApiResponseDto,
  ApiUnauthorizedResponse,
  ApiValidationErrorResponse,
  schemaRef,
} from '../openapi/api-responses.decorator';
import { ErrorResponseDto } from '../common/dto/error-response.dto';

function toClassroomWithMembers(classroom: {
  id: string;
  name: string;
  ownerId: string;
  createdAt: Date;
  updatedAt: Date;
  userClassrooms: Array<{ user: { id: string; name: string } }>;
}): ClassroomWithMembersDto {
  return {
    id: classroom.id,
    name: classroom.name,
    ownerId: classroom.ownerId,
    members: classroom.userClassrooms.map(({ user }) => user),
    createdAt: classroom.createdAt,
    updatedAt: classroom.updatedAt,
  };
}

@ApiTags('classrooms')
@ApiBearerAuth('bearerAuth')
@ApiExtraModels(
  ClassroomSummaryDto,
  ClassroomWithMembersDto,
  ClassroomMemberDto,
  TeacherSummaryDto,
  LastAnnouncementSummaryDto,
  CreateClassroomDto,
)
@Controller({
  path: 'classrooms',
  version: '1',
})
export class ClassroomsController {
  constructor(private readonly classroomsService: ClassroomsService) {}

  @UseGuards(JwtAuthGuard, RolesGuard)
  @Roles(Role.PROFESSOR)
  @Post()
  @ApiOperation({
    operationId: 'classrooms.create',
    summary: 'Criar turma',
    description: 'Requer role PROFESSOR; o criador se torna owner e membro.',
  })
  @ApiBody({ type: CreateClassroomDto })
  @ApiResponseDto(201, ClassroomWithMembersDto, 'Turma criada')
  @ApiValidationErrorResponse()
  @ApiUnauthorizedResponse()
  @ApiForbiddenResponse()
  @ApiConflictResponse()
  create(@Request() req: { user: AuthUser }, @Body() dto: CreateClassroomDto) {
    return this.classroomsService
      .create(req.user.id, dto.name)
      .then(toClassroomWithMembers);
  }

  @UseGuards(JwtAuthGuard)
  @Post(':id/join')
  @ApiOperation({
    operationId: 'classrooms.join',
    summary: 'Entrar em uma turma',
  })
  @ApiParam({
    name: 'id',
    type: 'string',
    format: 'uuid',
    description: 'ID da turma',
  })
  @ApiResponseDto(201, ClassroomWithMembersDto, 'Membership criada')
  @ApiValidationErrorResponse()
  @ApiUnauthorizedResponse()
  @ApiNotFoundResponse()
  join(@Param('id') id: string, @Request() req: { user: AuthUser }) {
    return this.classroomsService
      .join(req.user.id, id)
      .then(toClassroomWithMembers);
  }

  @UseGuards(JwtAuthGuard)
  @Post(':id/leave')
  @HttpCode(HttpStatus.OK)
  @ApiOperation({
    operationId: 'classrooms.leave',
    summary: 'Sair de uma turma',
    description:
      'Permitido somente a membro não-owner. O owner deve excluir a turma.',
  })
  @ApiParam({
    name: 'id',
    type: 'string',
    format: 'uuid',
    description: 'ID da turma',
  })
  @ApiResponseDto(200, ClassroomWithMembersDto, 'Membership removida')
  @ApiValidationErrorResponse()
  @ApiUnauthorizedResponse()
  @ApiNotFoundResponse()
  @ApiConflictResponse()
  leave(@Param('id') id: string, @Request() req: { user: AuthUser }) {
    return this.classroomsService
      .leave(req.user.id, id)
      .then(toClassroomWithMembers);
  }

  @UseGuards(JwtAuthGuard)
  @Get('my')
  @ApiOperation({
    operationId: 'classrooms.findMine',
    summary: 'Listar minhas turmas',
    description:
      'Retorna ownerId explícito para decisões de apresentação; a API continua autoridade de acesso.',
  })
  @ApiResponse({
    status: 200,
    description: 'Turmas das quais o usuário participa',
    schema: {
      type: 'array',
      items: schemaRef(ClassroomSummaryDto),
    },
  })
  @ApiUnauthorizedResponse()
  findMyClassrooms(@Request() req: { user: AuthUser }) {
    return this.classroomsService.findMyClassrooms(req.user.id);
  }

  @UseGuards(JwtAuthGuard)
  @Get()
  @ApiOperation({
    operationId: 'classrooms.findAvailable',
    summary: 'Listar turmas disponíveis',
    description: 'Exclui turmas das quais o usuário já participa.',
  })
  @ApiQuery({
    name: 'search',
    required: false,
    type: 'string',
    maxLength: 80,
    style: 'form',
    explode: true,
    description:
      'Pesquisa por substring literal do nome, sem diferenciar maiúsculas/minúsculas e preservando acentos. Parâmetro escalar único com serialização form; valores repetidos são rejeitados. O limite de 80 pontos de código Unicode é aplicado após remover espaços externos. Vazio significa sem filtro.',
  })
  @ApiResponse({
    status: 200,
    description: 'Turmas disponíveis',
    schema: {
      type: 'array',
      items: schemaRef(ClassroomSummaryDto),
    },
  })
  @ApiValidationErrorResponse()
  @ApiUnauthorizedResponse()
  findAvailable(
    @Request() req: { user: AuthUser },
    @Query() query: FindAvailableClassroomsQueryDto,
  ) {
    return this.classroomsService.findAvailableClassrooms(
      req.user.id,
      query.search as string | undefined,
    );
  }

  @UseGuards(JwtAuthGuard, RolesGuard)
  @Roles(Role.PROFESSOR)
  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({
    operationId: 'classrooms.delete',
    summary: 'Excluir turma própria',
    description:
      'Requer PROFESSOR owner. A primeira exclusão remove permanentemente memberships e announcements por cascata PostgreSQL e registra um receipt mínimo. Uma repetição reconhecida do mesmo owner também retorna 204 sem novo efeito.',
  })
  @ApiParam({
    name: 'id',
    type: 'string',
    format: 'uuid',
    description: 'ID da turma',
  })
  @ApiResponse({
    status: 204,
    description:
      'Turma e dependências excluídas permanentemente, ou repetição reconhecida do mesmo owner já concluída; resposta sem corpo',
  })
  @ApiValidationErrorResponse()
  @ApiUnauthorizedResponse()
  @ApiForbiddenResponse()
  @ApiExtraModels(ErrorResponseDto)
  @ApiResponse({
    status: 404,
    description:
      'Turma ausente sem receipt de exclusão correspondente ao usuário atual',
    schema: schemaRef(ErrorResponseDto),
  })
  delete(@Param('id') id: string, @Request() req: { user: AuthUser }) {
    return this.classroomsService.delete(req.user.id, id);
  }
}