# specs/001-app-quality-readiness/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Avisa Aí Professor API",
    "version": "1.0.0",
    "description": "Contrato de design da API REST v1. A referência executável será gerada pelos DTOs e decorators NestJS e ficará disponível somente fora de produção."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento local"
    }
  ],
...
  "paths": {
...
    "/api/v1/classrooms": {
      "get": {
        "operationId": "classrooms.findAvailable",
        "tags": ["classrooms"],
        "summary": "Listar turmas disponíveis",
        "description": "Exclui turmas das quais o usuário já participa.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "maxLength": 80 },
            "style": "form",
            "explode": true,
            "description": "Pesquisa por substring literal do nome, sem diferenciar maiúsculas/minúsculas e preservando acentos. Parâmetro escalar único com serialização form; valores repetidos são rejeitados. O limite de 80 pontos de código Unicode é aplicado após remover espaços externos. Vazio significa sem filtro."
          }
        ],
        "responses": {
          "200": {
            "description": "Turmas disponíveis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ClassroomSummary" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "operationId": "classrooms.create",
        "tags": ["classrooms"],
        "summary": "Criar turma",
        "description": "Requer role PROFESSOR; o criador se torna owner e membro.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateClassroomRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Turma criada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ClassroomWithMembers" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/classrooms/my": {
      "get": {
        "operationId": "classrooms.findMine",
        "tags": ["classrooms"],
        "summary": "Listar minhas turmas",
        "description": "Retorna ownerId explícito para decisões de apresentação; a API continua autoridade de acesso.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Turmas das quais o usuário participa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ClassroomSummary" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/classrooms/{id}/join": {
      "post": {
        "operationId": "classrooms.join",
        "tags": ["classrooms"],
        "summary": "Entrar em uma turma",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ClassroomId" }],
        "responses": {
          "201": {
            "description": "Membership criada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ClassroomWithMembers" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/classrooms/{id}/leave": {
      "post": {
        "operationId": "classrooms.leave",
        "tags": ["classrooms"],
        "summary": "Sair de uma turma",
        "description": "Permitido somente a membro não-owner. O owner deve excluir a turma.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ClassroomId" }],
        "responses": {
          "200": {
            "description": "Membership removida",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ClassroomWithMembers" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/classrooms/{id}": {
      "delete": {
        "operationId": "classrooms.delete",
        "tags": ["classrooms"],
        "summary": "Excluir turma própria",
        "description": "Requer PROFESSOR owner. A primeira exclusão remove permanentemente memberships e announcements por cascata PostgreSQL e registra um receipt mínimo. Uma repetição reconhecida do mesmo owner também retorna 204 sem novo efeito.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ClassroomId" }],
        "responses": {
          "204": { "description": "Turma e dependências excluídas permanentemente, ou repetição reconhecida do mesmo owner já concluída; resposta sem corpo" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": {
            "description": "Turma ausente sem receipt de exclusão correspondente ao usuário atual",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/announcements/classrooms/{classroomId}": {
      "get": {
        "operationId": "announcements.findByClassroom",
        "tags": ["announcements"],
        "summary": "Listar comunicados ativos de uma turma",
        "description": "Requer membership na turma.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ClassroomIdNamed" }],
        "responses": {
          "200": {
            "description": "Comunicados ativos da turma",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Announcement" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token de desenvolvimento. A referência não persiste nem fornece credenciais."
      }
    },
    "parameters": {
      "ClassroomId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" },
        "description": "ID da turma"
      },
      "ClassroomIdNamed": {
        "name": "classroomId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" },
        "description": "ID da turma"
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Entrada inválida ou campo não permitido",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Credencial ausente, inválida ou expirada",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "Role, ownership, autoria ou membership insuficiente",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Recurso inexistente ou não acessível",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Conflict": {
        "description": "Conflito de unicidade, limite ou lifecycle",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["statusCode", "message"],
        "properties": {
          "statusCode": { "type": "integer", "minimum": 400, "maximum": 599 },
          "message": {
            "oneOf": [
              { "type": "string" },
              { "type": "array", "items": { "type": "string" } }
            ]
          },
          "error": { "type": "string" }
        }
      },
...
      "TeacherSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "name"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" }
        }
      },
      "LastAnnouncementSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "title", "createdAt", "expiresAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      },
      "ClassroomSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "name", "ownerId", "teacher", "lastAnnouncement"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "minLength": 3, "maxLength": 80 },
          "ownerId": { "type": "string", "format": "uuid" },
          "teacher": { "$ref": "#/components/schemas/TeacherSummary" },
          "lastAnnouncement": {
            "allOf": [{ "$ref": "#/components/schemas/LastAnnouncementSummary" }],
            "nullable": true
          }
        }
      },
      "CreateClassroomRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 3, "maxLength": 80 }
        }
      },
      "ClassroomMember": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "name"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" }
        }
      },
      "ClassroomWithMembers": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "name", "ownerId", "members", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "ownerId": { "type": "string", "format": "uuid" },
          "members": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ClassroomMember" }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "AnnouncementAuthor": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "name"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" }
        }
      },
      "Announcement": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "classroomId", "title", "content", "expiresAt", "author", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "classroomId": { "type": "string", "format": "uuid" },
          "title": { "type": "string", "minLength": 3, "maxLength": 120 },
          "content": { "type": "string", "minLength": 3, "maxLength": 2000 },
          "expiresAt": { "type": "string", "format": "date-time" },
          "author": { "$ref": "#/components/schemas/AnnouncementAuthor" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}